Hoyong made AllTest view

+++ INTEX/Models/Billing.cs
@@ -26,12 +26,12 @@ namespace INTEX.Models
         public int? discountID { get; set; }
 
         [Display(Name = "Quote")]
-        public double? priceQuote { get; set; }
+        public decimal? priceQuote { get; set; }
 
         [Display(Name = "Price Before Discount")]
-        public double? priceBeforeDiscount { get; set; }
+        public decimal? priceBeforeDiscount { get; set; }
 
         [Display(Name = "Final Price")]
-        public double? finalPrice { get; set; }
+        public decimal? finalPrice { get; set; }
     }
 }
